// front/src/modules/transactions/interfaces/get-transaction.ts
import { GetAccount } from "@/modules/accounts/interfaces/get-account";

export interface GetTransaction {
  id: string;
  type: "debito" | "credito";
  amount: number;
  currentBalance: number;
  movement: string;
  accountId: string;
  account?: GetAccount;
  createdAtStr: string;
  status: boolean;
}